Error handling (business logic layer)

+++ api/logic.py
@@ -1,45 +1,67 @@
+from flask import abort
 import sqlalchemy as sa
+from sqlalchemy import or_
 
 from api import db
 from api.models import User, Task
 
 class TaskLogic:
+    
+    @staticmethod
+    def get_task(id, user):
+        """Get task by id"""
+        task = db.get_or_404(Task, id)
+        #regular user can only request their own task
+        if user.role != "admin" and not (user.id != task.created_by_id or user.id != task.assigned_to_id):
+            abort(403)
+        return task
+    
+    @staticmethod
+    def get_tasks(params, user):
+        """Get all tasks according to filters"""
+        #regular users can only request their own tasks
+        if user.role != "admin":
+            condition = [or_(Task.created_by_id == user.id, Task.assigned_to_id == user.id)]
+        else:
+            condition = [getattr(Task, key) == value for key, value in params.items()]
+        query = sa.select(Task).where(*condition)
+        print(query)
+        tasks = db.session.scalars(query).all()
+        return tasks
 
     @staticmethod
     def new_task(payload, user):
         """Create new task"""
+        #regular users cannot assign tasks to other users
+        if user.role != "admin":
+            payload.pop("assigned_to_it", None)
         payload["created_by_id"] = user.id
         task = Task(**payload)
         db.session.add(task)
         db.session.commit()
         return task
-    
-    @staticmethod
-    def get_task(id):
-        """Get task by id"""
-        return db.get_or_404(Task, id)
-    
-    @staticmethod
-    def get_tasks(params):
-        """Get all tasks according to filters"""
-        condition = [getattr(Task, key) == value for key, value in params.items()]
-        query = sa.select(Task).where(*condition)
-        tasks = db.session.scalars(query).all()
-        return tasks
 
     @staticmethod
-    def update_task(id, payload):
+    def update_task(id, payload, user):
         """Update task according to params"""
         task = db.get_or_404(Task, id)
+        if user.role != "admin":
+            if user.id != task.created_by_id:
+                abort(403)
+            else:
+                #regular users cannot assign tasks to other users
+                payload.pop("assigned_to_id", None)
         for key, value in payload.items():
             setattr(task, key, value)
         db.session.commit()
         return task
 
     @staticmethod
-    def delete_task(id):
+    def delete_task(id, user):
         """Delete task by id"""
         task = db.get_or_404(Task, id)
+        if user.role != "admin" and user.id != task.created_by_id:
+            abort(403)
         db.session.delete(task)
         db.session.commit()
